refactor: document shipped domains and harden shared layer boundaries
Move pagination and Cloudflare runtime behind feature-facing ports, document
weather/young-events/publications, raise unit coverage floors, and break
simple circular imports without changing public API shapes.
