+++ biome.json
@@ -49,9 +49,6 @@
         "noUnknownAtRules": "off",
         "noExplicitAny": "error"
       }
-    },
-    "domains": {
-      "react": "recommended"
     }
   },
   "assist": {
