+++ src/features/bus/components/bus-transit-map-types.ts
@@ -1,0 +1,8 @@
+export type Pos = { x: number; y: number };
+
+export type MapViewBox = {
+  height: number;
+  minX: number;
+  minY: number;
+  width: number;
+};
