+++ src/features/dashboard/server/compact-overview-read-model.ts
@@ -1,6 +1,6 @@
 import { loadOverviewTodoBundle } from "@/features/todos/server/todo-service";
 import { type AppLocale, DEFAULT_LOCALE } from "@/i18n/config";
-import { runCloudflareTraceSpan } from "@/lib/adapters/cloudflare-runtime";
+import { runCloudflareTraceSpan } from "@/lib/ports/runtime";
 import { withUserDbContext } from "@/lib/db/prisma";
 import { elapsedMs, monotonicNowMs } from "@/lib/log/observability-clock";
 import {
