+++ vitest.config.ts
@@ -12,10 +12,10 @@ export default defineConfig({
       include: ["src/**/*.ts"],
       exclude: ["src/generated/**"],
       thresholds: {
-        statements: 55,
-        branches: 48,
-        functions: 55,
-        lines: 57,
+        statements: 60,
+        branches: 53,
+        functions: 60,
+        lines: 62,
       },
     },
   },
