+++ worker-configuration.d.ts
@@ -1,5 +1,5 @@
 /* eslint-disable */
-// Generated by Wrangler by running `wrangler types --include-runtime=false` (hash: d2ad989fc7555a6e40bc574562bcb49e)
+// Generated by Wrangler by running `wrangler types --include-runtime=false` (hash: f97ba7e3377ad52e5202d7f1e4c90bb3)
 interface __BaseEnv_Env {
 	CALENDAR_EXPORTS: KVNamespace;
 	CATALOG_DETAIL_CORE: KVNamespace;
@@ -17,6 +17,13 @@ interface __BaseEnv_Env {
 	ASSETS: Fetcher;
 	NODE_ENV: "production";
 	APP_PUBLIC_ORIGIN: "https://life-ustc.tiankaima.dev";
+	DATABASE_URL: string;
+	ALLOW_DATABASE_SEED: string;
+	WEBHOOK_SECRET: string;
+	AUTH_SECRET: string;
+	AUTH_OIDC_ISSUER: string;
+	AUTH_OIDC_CLIENT_ID: string;
+	AUTH_OIDC_CLIENT_SECRET: string;
 }
 declare namespace Cloudflare {
 	interface GlobalProps {
@@ -29,5 +36,5 @@ type StringifyValues<EnvType extends Record<string, unknown>> = {
 	[Binding in keyof EnvType]: EnvType[Binding] extends string ? EnvType[Binding] : string;
 };
 declare namespace NodeJS {
-	interface ProcessEnv extends StringifyValues<Pick<Cloudflare.Env, "NODE_ENV" | "APP_PUBLIC_ORIGIN">> {}
+	interface ProcessEnv extends StringifyValues<Pick<Cloudflare.Env, "NODE_ENV" | "APP_PUBLIC_ORIGIN" | "DATABASE_URL" | "ALLOW_DATABASE_SEED" | "WEBHOOK_SECRET" | "AUTH_SECRET" | "AUTH_OIDC_ISSUER" | "AUTH_OIDC_CLIENT_ID" | "AUTH_OIDC_CLIENT_SECRET">> {}
 }
